validPeaksMatrix {MsCoreUtils}	R Documentation

Validation functions

Description

These functions are used to validate input arguments. In general they are just wrapper around their corresponding is* function with an error message.

Usage

validPeaksMatrix(x)

Arguments

x
object to test.

Details

validPeaksMatrix: see isPeaksMatrix.

Value

logical(1), TRUE if validation are successful otherwise an error is thrown.

Author(s)

Sebastian Gibb

See Also

Other helper functions for developers: between(), isPeaksMatrix(), rbindFill(), vapply1c()

Examples

try(validPeaksMatrix(1:2))
validPeaksMatrix(cbind(mz = 1:2, intensity = 1:2))

[Package MsCoreUtils version 1.4.0 Index]