Aggreagate quantitative features.

This function takes a matrix of quantitative features x and a factor (of length equal to nrow(x)) defining subsets, and applies a user-defined function to aggregate each subset into a vector of quantitative values.

User-defined functions must thus return a vector of length equal to ncol(x). Examples thereof are

medianPolish() to fits an additive model (two way decomposition) using Tukey's median polish_ procedure using stats::medpolish();

robustSummary() to calculate a robust aggregation using MASS::rlm();

base::colMeans() to use the mean of each column;

base::colSums() to use the sum of each column;

matrixStats::colMedians() to use the median of each column.

aggregate_by_vector(x, INDEX, FUN, ...)

Arguments

x
A matrix of mode numeric.

INDEX
A factor of length nrow(x).

FUN
A function to be applied to the subsets of x.

...
Additional arguments passed to FUN.

Value

A new matrix of dimensions ncol(x) and length(INDEX) with dimnames equal to colnames(x) and INDEX.

See also

Examples

x <- structure(c(10.3961935744407, 17.1663715212693, 14.1027587989326,
                12.850349037785, 10.6379251053134, 7.52885076885599,
                3.91816118984218, 11.1339832690524, 16.5321471730746,
                14.1787908569268, 11.9422579479634, 11.5154097311056,
                7.69906817878979, 3.97092153807337, 11.9394664781386,
                15.3791100898935, 14.2409281956285, 11.2106867261254,
                12.2958526883634, 9.00858488668671, 3.83120129974963,
                12.9033445520186, 14.375814954807, 14.1617803596661,
                10.1237981632645, 13.3390344671153, 9.75719265786117,
                3.81046169359919),
              .Dim = c(7L, 4L),
              .Dimnames = list(c("X1", "X27", "X41", "X47", "X52",
                                 "X53", "X55"),
                               c("iTRAQ4.114", "iTRAQ4.115",
                                 "iTRAQ4.116", "iTRAQ4.117")))
x

#>     iTRAQ4.114 iTRAQ4.115 iTRAQ4.116 iTRAQ4.117
#> X1   10.396194  11.133983  11.939466  12.903345
#> X27  17.166372  16.532147  15.379110  14.375815
#> X41  14.102759  14.178791  14.240928  14.161780
#> X47  12.850349  11.942258  11.210687  10.123798
#> X52  10.637925  11.515410  12.295853  13.339034
#> X53   7.528851   7.699068   9.008585   9.757193
#> X55   3.918161   3.970922   3.831201   3.810462

k <- factor(c("B", "E", "X", "E", "B", "B", "E"))

aggregate_by_vector(x, k, colMeans)

#>   iTRAQ4.114 iTRAQ4.115 iTRAQ4.116 iTRAQ4.117
#> B    9.52099   10.11615   11.08130  11.999857
#> E   11.31163   10.81511   10.14033   9.436692
#> X   14.10276   14.17879   14.24093  14.161780

aggregate_by_vector(x, k, robustSummary)

#>   iTRAQ4.114 iTRAQ4.115 iTRAQ4.116 iTRAQ4.117
#> B    9.52099   10.16203   11.08130  11.999857
#> E   11.45567   10.81511   10.14033   9.225928
#> X   14.10276   14.17879   14.24093  14.161780

aggregate_by_vector(x, k, medianPolish)

#>   iTRAQ4.114 iTRAQ4.115 iTRAQ4.116 iTRAQ4.117
#> B   10.39619   11.13398   11.93947   12.90334
#> E   12.85035   11.94226   11.21069   10.12380
#> X   14.10276   14.17879   14.24093   14.16178